+++ phpunit.xml
@@ -1,0 +1,7 @@
+<phpunit bootstrap="tests/bootstrap.php">
+    <testsuites>
+        <testsuite name="GameOfLifeTests">
+            <directory>tests</directory>
+        </testsuite>
+    </testsuites>
+</phpunit>
